+++ farmer/ncc/generators/__init__.py
@@ -1,4 +1,4 @@
 from .image_sequence import ImageSequence
 from .generate_with_mask import generate_with_mask
-from .data_loader import Dataloder
+from .dataloader import Dataloder
 from .dataset import ClassificationDataset, SegmentationDataset
